refactor: mv widget wrapper to components
